Enable Sentry Release tracking by default
If events have a 'release' attribute set, then Sentry will
automcatiically the first and last release in which events occured.

This change makes the client read the package's versionCode and use that
as the 'release' by default. The versionCode is an integer that
increases, and therefore is a perfect candidate for release tracking.

The user can continue to use an event capture listener to set the
release manaully, which will override this new default value.

Fixes #71